requirements.txt
python-dotenv==1.0.0
pyyaml==6.0.1
SpeechRecognition==3.10.1
pyttsx3==2.90
PyAudio==0.2.14
google-generativeai==0.3.2
openai==1.12.0
rich==13.7.0
psutil==5.9.8
requests==2.31.0
pytest==8.0.2